[Feat] Home Coordinator 프로토콜, 구현체 분리
